+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to this project will be documented in this file.
 
+## [Unreleased] - 2022-03-28
+- Added: [Entity finder] allow search for inserttags from event (added `ExtendEntityFinderEvent::addInserttag()`)
+- Fixed: [Entity finder] only one parent found for each entity
+
 ## [2.212.1] - 2022-03-28
 - Fixed: a potential error upon cache warmup and a missing system/tmp folder ([#43], [@qzminski])